Create a  Complex  heart Shape Using CSS and HTML.html
Transform the element on the screen to a heart. In the heart::after selector, change the background-color to pink and the border-radius to 50%.

Next, target the element with the class heart (just heart) and fill in the transform property. Use the rotate() function with -45 degrees. (rotate() works the same way that skewX() and skewY() do).

Finally, in the heart::before selector, set its content property to an empty string.